+++ packages/brick_graphql/lib/src/graphql_adapter.dart
@@ -38,12 +38,6 @@ abstract class GraphqlAdapter<_Model extends Model> implements Adapter<_Model> {
     ModelRepository<GraphqlModel>? repository,
   });
 
-  /// Returns an enum value based on its string name. For example, given `'POST'`,
-  /// and the enum `enum Methods { GET, POST, DELETE }`, `Methods.POST` is returned.
-  static T? enumValueFromName<T>(Iterable<T> enumValues, String enumName) {
-    return firstWhereOrNull(enumValues, (e) => e.toString().split('.').last == enumName);
-  }
-
   /// After Dart 2.12, `firstWhere` can be a non-nullable return value. Instead of an extension,
   /// a predictable static method is used instead.
   ///