+++ packages/brick_json_generators/lib/json_deserialize.dart
@@ -68,9 +68,8 @@ mixin JsonDeserialize<_Model extends Model, _Annotation extends FieldSerializabl
       // Iterable<enum>
       if (argTypeChecker.isEnum) {
         if (fieldAnnotation.enumAsString) {
-          final nullableSuffix = argTypeChecker.isNullable ? '' : '!';
           return '''$fieldValue.map(
-            (value) => ${providerName}Adapter.enumValueFromName(${SharedChecker.withoutNullability(argType)}.values, value)$nullableSuffix
+            ${SharedChecker.withoutNullability(argType)}.values.byName
           )$castIterable$defaultValue
           ''';
         } else {
@@ -115,8 +114,7 @@ mixin JsonDeserialize<_Model extends Model, _Annotation extends FieldSerializabl
       // enum
     } else if (checker.isEnum) {
       if (fieldAnnotation.enumAsString) {
-        final nullableSuffix = checker.isNullable ? '' : '!';
-        return '${providerName}Adapter.enumValueFromName(${SharedChecker.withoutNullability(field.type)}.values, $fieldValue)$nullableSuffix$defaultValue';
+        return '${SharedChecker.withoutNullability(field.type)}.values.byName($fieldValue)$defaultValue';
       } else {
         if (checker.isNullable) {
           return '$fieldValue is int ? ${SharedChecker.withoutNullability(field.type)}.values[$fieldValue as int] : null$defaultValue';